requirements.txt
ultralytics
numpy
opencv-python-headless
python-dotenv
fastapi
uvicorn
boto3
pymongo
pydantic
python-jose[cryptography]
concurrent-log-handler
passlib[bcrypt]
vidgear
bcrypt
python-multipart




